Refactored autocomplete
