+++ src/ui/components/common/autocomplete/index.js
@@ -1,0 +1,1 @@
+export { default as AsyncAutocomplete } from "./async";
